Add files via upload
No IPO chart on this one, but I did put a bunch of comments throughout. The IPO chart for input should consist of the choices that the user has to make for their weapon and their character. Processing would be to run the practice session or to play the potion minigame. Output would probably just be the game itself.