Plenty of edits
+ Modified some do-whiles
+ Added documentation for all functions
+ Added IPO chart

- Might have fucked up some code, but everything seems to work when tested
- Changed "\n" at the end of cout lines to endl - The professor seems to prefer endl

Definitely have a lot more edits to do in terms of how to output is presented to the user, and overall the way that some of the instructions and code is being presented. Some of the code is very hard to understand, so I might need to rearrange some stuff, and I'll definitely have to change some of the wording with the prompting.