Added functions and necessary documentation
This code should be able to do the exact same stuff as V7 of the program, but the readability of this code is much better. Still need to edit the output to make it look nicer (using setw), but overall, it seems like the program is complete. Be sure to look at the instructions to see if I missed anything (like with the specifications of what types of functions to use and if I used them all).